sdk/python/pyproject.toml
[project]
name = "graft-py"
version = "0.1.0"
description = "GRAFT protocol SDK for Python — host-side hardware control"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = []

[project.optional-dependencies]
serial = ["pyserial-asyncio>=0.6"]
mqtt = ["aiomqtt>=2.0"]
all = ["pyserial-asyncio>=0.6", "aiomqtt>=2.0"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
strict = true
python_version = "3.10"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
